+++ TacViewDataLogger.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28922.388
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TacViewDataLogger", "TacViewDataLogger\TacViewDataLogger.csproj", "{BBF23630-1F98-45E0-8DA8-1108774217D6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BBF23630-1F98-45E0-8DA8-1108774217D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BBF23630-1F98-45E0-8DA8-1108774217D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BBF23630-1F98-45E0-8DA8-1108774217D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BBF23630-1F98-45E0-8DA8-1108774217D6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {190E1EC4-2E3B-492F-B3A4-37E65461DD32}
+	EndGlobalSection
+EndGlobal
